+++ modules/spring-core/src/main/kotlin/com/esprito/spring/core/inspections/SpringConfigurationPropertiesNullableParametersInspection.kt
@@ -1,21 +1,22 @@
 package com.esprito.spring.core.inspections
 
+import com.esprito.inspection.SpringBaseUastLocalInspectionTool
 import com.esprito.spring.core.SpringCoreBundle
 import com.esprito.spring.core.SpringCoreClasses
 import com.esprito.util.EspritoPsiUtil.getHighlightRange
 import com.esprito.util.EspritoPsiUtil.isMetaAnnotatedBy
-import com.intellij.codeInspection.AbstractBaseUastLocalInspectionTool
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.PsiElementVisitor
 import com.intellij.uast.UastVisitorAdapter
+import org.jetbrains.kotlin.idea.KotlinLanguage
 import org.jetbrains.kotlin.psi.KtNullableType
 import org.jetbrains.kotlin.psi.KtParameter
 import org.jetbrains.uast.UParameter
 import org.jetbrains.uast.getContainingUClass
 import org.jetbrains.uast.visitor.AbstractUastNonRecursiveVisitor
 
-class SpringConfigurationPropertiesNullableParametersInspection : AbstractBaseUastLocalInspectionTool() {
+class SpringConfigurationPropertiesNullableParametersInspection : SpringBaseUastLocalInspectionTool() {
 
     override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor {
         return UastVisitorAdapter(ConstructorParameterVisitor(holder, isOnTheFly), true)
@@ -26,9 +27,10 @@ class SpringConfigurationPropertiesNullableParametersInspection : AbstractBaseUa
     ) : AbstractUastNonRecursiveVisitor() {
 
         override fun visitParameter(node: UParameter): Boolean {
-            if (node.lang.id != "kotlin") return true
+            if (node.lang != KotlinLanguage.INSTANCE) return true
             if (node.isFinal) return true
             val ktParameter = node.sourcePsi as? KtParameter ?: return true
+            if (ktParameter.hasDefaultValue()) return true
             if (node.getContainingUClass()?.javaPsi
                     ?.isMetaAnnotatedBy(SpringCoreClasses.CONFIGURATION_PROPERTIES) != true
             ) return true