+++ modules/spring-core/src/main/kotlin/com/esprito/spring/core/inspections/SpringProfileWrongPlacementInspection.kt
@@ -1,5 +1,6 @@
 package com.esprito.spring.core.inspections
 
+import com.esprito.inspection.SpringBaseUastLocalInspectionTool
 import com.esprito.spring.core.SpringCoreBundle
 import com.esprito.spring.core.SpringCoreClasses.BEAN
 import com.esprito.spring.core.SpringCoreClasses.COMPONENT
@@ -8,7 +9,6 @@ import com.esprito.spring.core.service.SpringSearchService
 import com.esprito.util.EspritoPsiUtil.getHighlightRange
 import com.esprito.util.EspritoPsiUtil.isMetaAnnotatedBy
 import com.esprito.util.EspritoPsiUtil.toSourcePsi
-import com.intellij.codeInspection.AbstractBaseUastLocalInspectionTool
 import com.intellij.codeInspection.InspectionManager
 import com.intellij.codeInspection.ProblemDescriptor
 import com.intellij.codeInspection.ProblemHighlightType
@@ -18,7 +18,7 @@ import org.jetbrains.uast.UClass
 import org.jetbrains.uast.UMethod
 
 
-class SpringProfileWrongPlacementInspection : AbstractBaseUastLocalInspectionTool() {
+class SpringProfileWrongPlacementInspection : SpringBaseUastLocalInspectionTool() {
 
     override fun checkMethod(
         uMethod: UMethod,