+++ modules/spring-core/src/main/kotlin/com/esprito/spring/core/inspections/SpringUnknownAliasMethodInspection.kt
@@ -1,5 +1,6 @@
 package com.esprito.spring.core.inspections
 
+import com.esprito.inspection.SpringBaseUastLocalInspectionTool
 import com.esprito.spring.core.SpringCoreBundle
 import com.esprito.spring.core.SpringCoreClasses
 import com.esprito.spring.core.service.AliasUtils
@@ -8,7 +9,6 @@ import com.esprito.util.EspritoPsiUtil.getHighlightRange
 import com.esprito.util.EspritoPsiUtil.isMetaAnnotatedBy
 import com.esprito.util.EspritoPsiUtil.toSourcePsi
 import com.intellij.codeInsight.AnnotationUtil
-import com.intellij.codeInspection.AbstractBaseUastLocalInspectionTool
 import com.intellij.codeInspection.InspectionManager
 import com.intellij.codeInspection.ProblemDescriptor
 import com.intellij.codeInspection.ProblemHighlightType
@@ -18,7 +18,7 @@ import com.intellij.psi.util.parentOfType
 import org.jetbrains.uast.UMethod
 
 
-class SpringUnknownAliasMethodInspection : AbstractBaseUastLocalInspectionTool() {
+class SpringUnknownAliasMethodInspection : SpringBaseUastLocalInspectionTool() {
 
     override fun checkMethod(
         uMethod: UMethod,