+++ modules/spring-core/src/main/kotlin/com/esprito/spring/core/inspections/quickfix/AddQualifierQuickFix.kt
@@ -7,7 +7,9 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.command.WriteCommandAction
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
-import com.intellij.psi.*
+import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiFile
+import com.intellij.psi.PsiModifierListOwner
 import com.intellij.psi.codeStyle.JavaCodeStyleManager